ANS2014 Changes Only
I created a new .gnu file in ASN2014, but changed nothing in the actual
code.  It may be better if I save all of this with the actual paper and
get it out of my research code.  I will probably do that later.  This
commit is essentially only to save this new .gnu file and its output and
does not affect the actual code at all.