Progress: Believe I have finished updating pdf plots through re-write
The atomic mix and LP solves do not have realizations.  Therefore, unless you want to make batches in some way, pdfs of leakage values are meaningless.  With 'batches' of 1 particle all counts would either be '0' or '1'.  Alternatively you could plot the average value for all of the particles.  This would actually be potentially useful information, and if desired, one can do this by hand.  In the future I would read bin location values in from another file, put the one contribution into the appropriate bin, then print the new file whos first column (bin location) is the same as the file read from, and the second other columns are a bunch of zeros and one non-zero term.  This would be a lot of work for very little benefit however.  Once again, if desired for a paper or analysis, a file like this with an appropriate gnu can be created by hand without too much hassle.

This commit renders the previous commit (b5bddfd) moot.  When I learn how to delete old commits, this may be a good place to practice.