Progerss: Allow different scatrats
I believe I have fixed this!  See notes from 10-23-14 or right after.
In 'material' mode, different xi values, so that summation is normalized... multiplying Co term by it gives shape.  I was not always using the right option for 'KLrxi_point', effectively using the 'totxs' solve when I was in 'material' mode for all except the scattering ratio.  I needed to pass 'total  ' when running in 'material' mode, so that the function would distinguish 'material' mode versus 'totxs' mode for most of its calls.
The variance seems low on final leakage results in some cases, but that may just be part of the system.
Next I need to calculate 'meanadjust' for each mode, or scale by variance, or something, but it needs to be dealt with in this new framework somehow.  Then it appears this method is up and running.