Progress: 'material' mode bug fix
In 'KLresearch', in 'KL_collect', where xi distributions are made,
changed integration scheme to make larger cross section correspond
with the positive xi values.
Before this values were good half of the time, but not the other.

Variances always seem to be a little low with this method.
Transport results seem to be pretty accurate if there is not much
negativity in the problem, as before, but a new negativity issue
has been introduced.  If the scattering or absorbing cross section
is very negative, the problem seems to be thrown off as well.