Add negwgts: Added gnu to plot negwgt cases
Added folder 'auxiliary', created for tools that are not directly part of the code
but are part of the research going into the code.
Added 'negwgt.gnu' in that folder, which is a gnu that can be used to look
at the different cases for choosing sigma_t^* in the negwgts implementation.