Add negwgts: 2 options only, added sigs array
Reduced to only two options for choosing refsig, user and adaptive.
For adaptive, allowed input for maximum acceptable increase, and
for number of bins to collect sigs and sigt.
These will be the bins over which a woodcock method chooses where to
either perform or not perform a weight adjustment based upon the local
refsig.  Each bin will have a refsig based upon the largest refsig
required in that bin from the samples of the edges and center.