Add negwgts: Finished adaptive implementation
I think it is working properly.
Results appear to be no more or less stable for the adaptive method
than the non-adaptive negwgts method.  If anything the non-adaptive
is more stable.  The new is quicker.
Brian Franke has suggested that the problem may simply be unstable,
which would mean any implementation of this method would be unstable.
That may be the case.

After roughly 2% of all sampled cross sections being negative the
method begins to be unstable.  This is a rough boundary, as I have seen
cases that break this mold in both directions, but it seems about correct.