Add negwgts: Improved adaptive woodcock, still not good
Added option for both adaptive methods to take more samples per
bin for calculating maximum refsig.
Also, for woodcock method, allowed refsig to be calculated on the fly
adjusting for any that are above the woodcock ceiling by reducing
the refsig to the woodcock ceiling.
Also, with this method, there was a bug, the refsig for the bin
was calculated based on the old position, not the new one.
Even with the bug fix, finer ceiling samples, and on-the-fly refsig
calculation, stability is only minimally improved. It was worth a shot.