rng: Updated rng handling
Began using mcnp_random's initialization and advancing subroutines.
Specifically, each unique operation which requires random numbers
(binary and KL material generation, and each MC method) was assigned
an application number.  In the new module 'rngvars' a stride can be
defined.  The number generator advances a number of particle histories
equal to this stride times the app number and anything else entered,
like realization number or particle history number.  In 'rngvars',
there is a simple subroutine which sets the application number so that
the rng will be advanced the correct amount for this application.

The program runs and gives believable results.  The best way to test this
will be when the code runs in parallel; the transport results should
be independent upon the order the operations were performed in, and
dependent only on defined parameters like realization number, particle
number, stride, seed, etc.  In other words, multiple parallel runs should
always yeild the exact same answer even though the job is split up and the
order of evaluation will not be the same every time.