cleanup: Cleaned up printing files
Made all plots and plotting files in a specific folder.
Got rid of old plots in 'plots'.
Ensured plots and plot files not sticking around in main folder.
Ensure trying to rm or mv files that do not exist will test instead
of throwing an error as before.  This cleans up what is printed out.