MLMC: added various solution arrays
Added:
M_optsamps - optimal number of samples as of this calculation and last
uflux - response function u, for our application it will probably always be flux
Q_ufunctional - a functional of u at each Level, can be with a norm of u or say leakage value
G_ufunctional - difference between Levels of Q_ufunctional, used in MLMC algorithm
Gave - average of G_ufunctional values for each Level
Gvar - variance of G_ufunctional values for each Level

Each of these has been allocated, initialized, deallocated, and enlarged
in Level (and number of cells for uflux) and number of samples.
Each of these five operations has a designated subprogram to perform this
operation.
When new values are added, they are initialized and if possible the desired
value is fed into them (for M_optsamps and numMLMCcells).