MLMC: wrote flux solve driver
Wrote flux solve driver which runs FEmain.
Modified FEDiffSn to deallocate arrays, arrays not interested in externally
simply at the end of the invocation of FEmain, and arrays interested in
by external parties in a differen deallocate array that the external subprogram
can call.  This allows the collection of flux values from an invocation.

Also added a little more infrastructure to NewSample solver, such as subroutine
for creating input files for the FE solver to solve.  This subroutine has pseudocode
in it, but no real code yet.

Because of the timing of this commit, I have not tested to make sure all of my
additions work.