MLMC: added input parser and passer
Added code that will take the valid entries for xs definitions,
transform them into the form needed by my FE solver, and
pass those variables along with the number of cells.
The FE solver also got a boost with a flag that is true when
this info is passed from the greater code (so that it knows not
to read those values from the input file), and false when it
should just read those values from the input file.