MLMC: begun opt num of samples estimate 2
Fixed bug: was setting old number to new number twice, once in new
code chunk and in the wrong place.
I think code is doing what it is expected to do with two exceptions:
1) Need to add error estimate calculation to decide if done iterating
2) I think QoI should be taken from multiple cells past the first
   level that all fall within space of cell in level 0.

Also, I have added print and read statements here to allow stepping
through the execution of the code, to help in ensuring correct
running and debugging.