MLMC: functional evaluation fixed (I think)
I think functionals should be taken over average values from
all cells in levels greater than 0 which physically align
will cells from level 0.  In other words, for a Beta of 3,
the value the functional taken for the first three cells
of level 1 should be over the average of the values in these
three cells.  This value will mimic using the values in only
the first cell in level 0.  Since the functional I am using
thus far is the L2 norm of the flux, I average the first three
cells (or 9 for level 2, or 27 for level 3), then square that
value.  These squares are taken throughout the mesh, then
the square root is taken of that quantity to yield the functional
output which is compared from different levels.