MLMC: added M_optsamps tally and fixed bug
First: Added another column to M_optsamps which keeps track
of what the most informed estimate of the optimal number of
samples was.  This will allow knowledge of how many 'extra' were
run that didn't need to be.

Second: When calculating the optimal number, the second term
is the sqrt of Gvar/Gave.  Sometimes this is trying to take the
sqrt of a negative number.  I am now taking the sqrt of the absolute
value.  I think this is perfectly valid, though I have found it
nowhere in the published or presented theory.