MLMC: fixed two bugs, developed more
Lethal bug: Due to a transcription error in my notes,
I was calculating the optimal number of samples wrong, using
the Gave in a denominator when I should have used a work term.

Non-lethal bug: Since using 'MLMCevalNewSamps' in two places
I needed to change how the lower bound of new samples to be
evaluated was determined with updated logic.  The detriment
from this was simply re-computing samples.  This did give a
chance to perform a sanity check on the rng usage, however,
as duplicate samples yielded the exact same flux functional
results.

I also added a flag that can be turned true or false to for the
'read' statements, and changed 'qq' in the total error estimate to
2 from 1, which I think might be more appropriate here.

The code does not perform as expected though, as it seems to take
less samples and the Monte Carlo error, error 2, is dominated by
level 0 contributions, and much higher than I think it should be.