MLMC: options with functionals part 5
I did a lot here.  Hopefully I can remember it all.  Now...
Only one functional is tested to converge.  The user sets which one
by making that one have a '1' in the field for whether it converges.
I added tests for what values are input here with comments.
The estimated error of each functional is printed, along with the
original tolerance, and either 'converged' or 'not converged'.
The contributing errors are also printed.
Near the end output the number of samples calculated and the
optimal number of samples is printed.  The first is as large
as or larger than the second.
Other things were likely done with the functionals, but I can't remember
them.