MLMC: options with functionals part 7
Added .gnu files in 'auxiliary' which can plot the outputs from the functionals
of the flux in the code.  In many cases this does not make much sense,
but can be useful for looking at the flux profile if the L1 or centered
option was taken in each cell or evenly spaced groupings of cells.
This will allow the expected value of these functionals along with the
estimated error upon that value to be compared when solved with other means.
In other words, the L1 or centered values over each cell or evenly spaced
groupings of cells can be compared against a benchmark method.
These gnu files must be used manually though, they are not incorporated
in the main code.