MLMC: spatial study pt2-solve functs
Here a subroutine has been developed using some previously
established subprograms, but also with a little time-saving
allocation and initialization in the subroutine itself,
which solves at average values for the chosen geometry and
material setup for functionals chosen.  It solves to each
level up to the user defined max level.

Also, the diffusion results had been used wrongly here
and in MLMC before.  Apparently the correct usage for
those diffusion values is not cell-centered, but cell
edges?  I don't remember this, but this seems to be
the way it is used.