MLMC: spatial study pt4-plots
Added gnu file in auxiliary/MLMCfuncts which plots data
printed in plots/MLMCfuncts/spatialconv.out.  Currently data is
present for diffusion solve of basic problem with a collection
of functionals, L1, L2, and center based.  The first two
have a log-log slope of about -2, the last of about -1.
This run used a spatial refinement factor of 3.