MLMC: spatial study pt5-slopes
Added to the error and printing subroutine a section of code
which uses the errors and calculates the slope from each possible
set of points, printing all of these in an organized fashion
to plots/MLMCfuncts/spatialslopes.out.

For diffusion problems the observed convergence was about
-2 for L1
-2 for L2
-1 for center based

For Sn and Sn with DSA problems the observed convergence was about
-3 for L1
-2 for L2
-1 for center based

Each is regardless of chosen refinement size.