MLMC: iter study pt2-structure and iter control
Left in a state which includes comments to showcase how this works,
this commit has introduced three variables inside FEDiffSn.
'flconvstudy' is whether or not an iterative convergence study is underway.
'max_iter' is how many iterations should be solved this time in order
  to get functional data at this iteration level.
'flnewiter' lets the convergence study loop know when to stop (when the
  solve has converged).
It would be more computationally efficient to get the information needed
at each iteration and only solve once, but these solves are cheap enough
that I just re-solve up to one more iteration than last time each time.

I also changed 'samplespatialInput' to 'sampleconvInput' so that this
input parser could be used in both cases with an appropriate name.

The program can now run the solver to successively higher numbers of
iterations in order to collect functional data at each iteration.
The first iteration is 0.  This captures the initial flux, not a solved
flux, and was mostly out of convenience since the 'Q_ufunctional' variable
is already solved for what is usually ilevel=0, but will here be iiter=0.
I don't think this data will be used in any way later, I did this this way
just to make things easier by fitting the already established form.