MLMC: iter study pt2.2-DSA and demonstrationless
In the previous commit I implemented iteration stopping mechanisms
for the Sn method, and forgot to for the Sn with DSA implementation.
I did that here, and took out things only added to demonstrate that
my iteration control methods work, so that the code is ready for the
next step.