MLMC: iter study pt3.1-collect functionals
I:
1) got away from using icase equal to a number and went to what I
should be using, which is MLMCcases(icase) equal to a word.
2) carried icase into solveSamples in each place it is called.
3) allowed ilevel in solveSamples to be the maximum iteration
desired this time for the iterative convergence study.  In this case
iilevel, used for some of the functionality of ilevel, is
set to 0, since the iterative convergence study will always
take place at one specified spatial refinement level, which
I have labelled 0.

These changes allow me to re-use 'solveSamples' with little
intrusive implementation for the iter conv study where I am
not refining spatially, but have instead used that rank to
signify iterative refinement.