MLMC: benchmark pt4-calc err and errSEM
In error calculation and printing subroutine, calculated
discretization error as compared with most converged level
and calculated SEM at chosen CI for this err by propagating
the SEMs at ilevel and the final level through the error
formula.

I also took a rank out of all instances of err_ufunctional
and updated how the code handles if it will allocate.