Quadrature: added netlib routine
My quadrature technique has been an executable which
produced an output file for the nodes and another for
the weights.  I have for some time wanted to find or
write a good routine to solve quadrature weights and
nodes without extra files.  I found one at
netlib.org/go/gaussq.f.  It was in fortran 77, however,
and required a seperate file to handle double point
precision.  I modified to run here, basically ignored
the picky (but possibly useful) precision file,
and made this routine part of 'utilities.f90'.
Currently it is in its test state at the beginning
of my main.  I next need to use this subroutine
to replace my old, temp-file-dependent method.
This is convenient timing to have found this subroutine
since the other method just started giving me a
headache, not wanting to give me permissions to my
temp files.