Gauss Proc: Large, Gauss-based
I believe Gaussian-random process through KL implemented and working.
Plots of internal flux and leakage pdfs are similar to same Adams
problems, especially if a problem with constant scattering ratio.

Sampling uses same samples, or correlated random number for
geometry sampling and MC sampling.  This helps answers be similar.
I should make this an option.

Input used is that derived from Markov-based case.  This is not bad,
but need to be able to use input designated for this process.

So it is not complete, but I think what is there is right.