Gauss Proc: Gauss-based w/ Mark or on own
This commit allows input for Gauss-based to be different, or stand
alone from Markov input.  There are also a few bug fixes, but
I'm not sure what all went into this, as I fiddled with several things.
I did fix usage of odd numbers of eigenmodes with pdf-based sampling.
I did re-solve the eivenvalue problem when using Gauss-based input.
I did pull the eigenvalue solver and the reconstructor into the outermost
MC algorithm.
I did load Gauss-based values at the beginning of the outermost MC algorithm,
in MCallocate.
I think this is all that I did, but I'm not sure, so if more information is
needed use git diff.

Results are very promising.