Gauss Proc: Neg stats variable change2
This time changed 'negcnt' to its compliment 'numPosRealz'.
I moved the variable from being a KLvar to a genRealzvar.
With the new method I will need to normalize according to
realizations not 'thrown out', and thus positive.  It made
more sense to keep the data in the positive form, and both forms
would be unnecessary.
I also initialized this in MCallocate.  I don't think negcnt
had been initialized.