Mean Adjust: KLrxi_point corrected and beefed up
KLrxi_point is the function which when called returns the value
of a cross section at a point x in the domain.
It can return a value for total, scattering, or absorption cross sections.

I had been calculating scattering and absorption cross sections wrong
when using meanadjust.  'meanfrac' had been spliting the contribution
of scattering and absorption in the fictitious meanadjust cross section,
but it was splitting in correctly.
The function is now written in a way that avoids doing this, though
I have solved out how it could be done.  Instead I solve total cross
section and a scattering or absorption ratio.  This requires a
recursive use of the function.  In some recursive calls I turn a flag
off an then back on.  I believe the function works as it should.

I also changed chxstype from a character string to length 7 to
an arbitrary length.