Neg Stats: New approach, large commit
In this commit I seek to consolidate 1. negsearch, 2. local max search,
and 3. zeros search into one master subroutine which performs any
of these desired with greater accuracy than previous routines.

I have added subroutine KLr_analyzerealz which accepts purpose flags
of which goals it is to accomplish.  It successively refines a grid
of KL derivative evaluations until all bounds on extrema are located.
It searches in these bounds and finds the extrema.  It tests each for
being a max or min and positive or negative.  If any are negative
the realization contains negativity.  All maxima are stored for use
in WMC.  Between successive extrema in which one is positive and
the other negative the zero is found.

Beyond the initial grid refinement, all searched are binary.

A function was added to 'utilities' which finds the sum of an
arithmatic series.  This was helpful in allocating maximum possible
required matrix lengths based on the number of eigenvalues.

This new function in its current state replaces 'KLr_negsearch'
entirely.  It has collected the information to replace searches
in setting WMC ceilings and finding meanadjust zeros, but these
replacements have not been made.  The zeros could also be utilized
to collect neg stats data which is more useful than the current data,
which is collected based on what particles experience, not on the
actual realizations.

I believe everything works properly.  A max same iteration value of
3 was the lowest value sufficient in all convergence studies run, though
it is possible some cases require a larger number.  It would be good
to add provision for a missed negativity, local max, or set of zeros.