redosetzero pt4: KLr_point modified to handle derivatives
I included an additional optional argument in KLr_point
which is the order of derivative desired.  The default is 0.
I added a small subroutine which loads mean values for order 0
or 0 for higher orders, which is what these terms are for
an order of derivative.  I updated the eigenfunction term
evaluater based on order.  Every fourth order is the same
except that there are more alphas.

This passes a benchmark test.  I need to utilize higher
orders instead of KLr_derivative and KLr_concavity.  This
should be the next commit.