redosetzero pt5: KLr_analyzerealz to KLr_realznegandzeros
I added function 'geometricsum' in utilities.

I added flag 'flfindzeros', such that true searches for zeros of
function and not only determines whether the realization is positive or
negative.

I changed KLr_analyzerealz to KLr_realznegandzeros.  This involved an
entirely new algorithm.  This algorithm breaks a realization for a cross section
up into segments, and first searches for bounds on zeros in each segment
then finds the zeros.  If flfindzeros is false whenever any negative value
is sampled, the subroutine kicks out flrealzneg as true, exiting the routine
immediately.  Any negativity should be found in finding bounds of zeros.  If
flfindzeros is true, the routine continues finding zeros one part of the routine
at a time.  When finding zeros this routine will be used twice,
once for scattering and once for absorption cross sections.

I changed KLr_findnextextremabounds to KLr_refinezerogrid.  This also
involved enough changes that the algorithm is essentially new.  This algorithm
uses function evaluations to find bounds on all zeros in the realization
segment passed to it.  It can be used for any defined cross section type
and for any order of derivative of the KL expansion in a realization.
Next will come a subprogram which uses those bounds to find the actual zero.

This commit does well in benchmark testing.  I believe it is correct.
Negative realizations are chosen if either the scattering or absorption
cross section is negative.  In KLr_reconstruct.f90 around line 152, I have
used KLr_negsearch instead of new routines to confirm that the old routine gets
the same number of negative realizations.  I used KLr_negsearch for abs and scat
separately, and removed the Heaviside capability of line 738 to use the total
cross section.  Either method yielded the same number of negative realizations,
at least when using 4 segments of refinement 6 each.

I am committing after not looking at this for a few weeks.
I think I am on track, but I may have missed something.  It appears that I was
trying to debug something in the routine around
line 533 of KLreconstruct.f90. --It appears that I was trying to exit the immediate
and outer do-loop using exit statements.  I have done this now by labeling
the do-loop and exiting the loop with that label.  I will continue to
search for anything else that looks suspicious.