redosetzero pt5.4: clean and moth KLr_r...
I got rid of old code, and effectively mothballed KLr_realznegandzeros,
at least in determining whether negative or not.  The old routine is
just as accurate and much faster.

I may still use this new routine as part of meanadjust, but the old may
be sufficient and faster, so I may not.  As of now, the new shiny routine
is not used.