redosetzero finish
I tidied up a few loose ends and benchmarked.  The new handling
of negative cross sections as zero and the related meanadjust
does as well as or better than before.

As a recap, before I had set the total cross section to zero if
it was negative, and alloted sigs and siga the ratio of the
local magnitudes.  I realized this is not faithful to the physics,
so now I set scattering and absorbing cross sections equal to zero
independently, and sample the total cross section by summing them.

Similarly, meanadjustments are performed for sigs and siga independently
instead of on sigt.

With a quick survey of Adamscase problems, the new methods do much
better than before, in some cases considerably better than LP, though
in others still considerably worse.  I should quantify this.  If I can
predict which cases this method will do better, it may still be worth
further pursuing.  If Dr. Prinja thinks it advisable, I may perform
variance preservation as well.

In my last commit I ignored klmeadadjust.mod and the code would not
compile.  I should spend some time and make sure the dependencies
are in line where I can compile from scratch.