Gauss Proc: sigs and siga xs diff rand
The Gaussian process should use different random numbers to sample
values for the scattering and absorption cross section.  Up until now
both have used the same random number as in the Markovian-based approach.
Now the user has the option for either.

This appears to be implemented right, though a careful look of some kind
for further verification would be nice.