LN Proc: First implementation
I think I have implemented a log-normal process correctly.
And by correctly I mean imitating what Erin Fichtl did with
Dr. Prinja in the 2011 JQSRT paper, which is to assume
an exponential covariance for the Gaussian process and
use the log-normal correlation length.  Leakage values
and realization plots seem reasonable, but I need to
compare flux profile with the paper and search for anything
else off.

Next I plan to curve fit a correlation length using an
exponential covariance for the underlying Gaussian.
This should make the method a little more accurate.