LN Proc: fexpfit.py
Added python script 'fexpfit.py' which reads 's', 'a', and 'lamcsig'
and returns 'lamcw'.  This takes the log-normal correlation length,
plots the related Gaussian correlation function, assumes that this
function is exponential, and finds the best fit correlation length
for the Gaussian correlation function.  This is not exact, but better
than using the log-normal correlation length.

Also added a function in 'utilities.f90' to print s, a, and lamcsig
to a text file, run the script, and collected lamcw from a text file
which the .py script creates.

Currently this is tested in the main, 'stochastic.f90', but this will
not be part of the commit stream after this.