LN Proc: testing dist pt2
Changed store_in_bins in 'utilities' some.  This is used elsewhere,
I believe now is much better.

Printed data for ave and var and pdf.

Plotted using .gnu files for ave and var and for pdf.

Corrected a couple variables -bugs- used in this process.

Added option for 'noplot', 'preview', or 'plot'.

This routine can be used with Gaussian or Log-normal implementations,
though the routine and many variables are named after Log-normal 'LN'.

This routine can be used for 'totaln', 'scatter', or 'absorb' xss.