LN Proc: Corr plots and Var
I fixed how the variance was split up.

I made correlation plots either plot for MB or GB with right
inputs.  Eigenvalues seemed off before for GB, but I cannot
replicate.

As part of the correlation invocation, I removed determination
of CoExp from this subroutine in favor of global_allocate,
and removed internal if-logic, instead placing it outside
the function call so that I can use different if-logic
for the two invocation places.