LN Proc: Var test pass
KL_Cochart was dependent on the variance.  It is now simply
normalized to approach 1.

It is invoked in two places now, one only when using MB input,
one only using GB input.

I also no longer need to create a comparison between the
Expected and Actual variance.  This was redundant in the
first place, and doubly so now that this plot is produced
without multiplying by the variance.