Simplify: rm fixed point
I had investigated a 'fixed point' usage of the KL expansion,
but this is not how it should be used.  It should be used in
the way I previously labeled 'fixed xi'.  Not that is just,
'using KL'.  I have attempted to remove all remnants of the
'fixed point' approach.

I have also renamed things including 'KLrr' to 'KLr'.
'KLrr' stood for KL reconstructed realizations, but we are
interested in KL now not only when reconstructing, but also
for simply constructing.  This redundancy in 'r's is no longer
accurate.  Besides, it was an extra character of clutter.