Simplify: Only 1 case pt6--rm flmode
In the subroutine 'genReal', flmode was a parameter to
tell the subroutine which type of realization operation
to perform.  It has been replaced by chTrantype.

One call is made to plot realizations when transport may
not be desired.  Maybe I could split this subroutine into
three subroutines--genbinaryReal, genLPreal, and genatmixReal--
That would remove the need for 'flmode' or its replacement
altogether and be more clear.