Simplify: time
I removed nearly all time tracking.
Time tracking now consists of an initialization of 't1',
then calls to compute progress on an iterative task, like
transport over realizations.  Time is no longer part of the
output printed to a file or the screen.  If you are
interested in the total time, run as 'time ./astochastic'.
If you are interested in what takes up the most time, use
the profiling tools, choose input carefully to test, or modify
code to isolate the code sections of interest.