Simplify: rm disthold
This was a 'workaround' that I had implemented for
handling negative cross sections.  It was not rigorous--
only a game.  It produces some improvement in results in
cases I looked at, but Dr. Prinja doesn't like it since
it is just a game and not rigorous... and I agree: I
wouldn't want to publish that.  So it is now gone.