Simplify: main organization
Hmm, this one kind of got away from me: I don't remember
everything I did in this commit.  The main idea was that I
tried to organize the 'main', which is 'stochastic.f90'.
I tried to make dependencies for subroutines show here instead
of internally, and make the structure as simple as possible.
I removed 'KLres' and 'KLrec', as better logic accomplishes what
they did with other variables.