Simplify: init and alloc pt1--plot test
I added a test for each plotting option to ensure
that either 'noplot', 'plot', or 'preview' was entered
by the user.

I then reorganized some logic to save space and complexity,
relying on the necessity of one of these three being the value
of the option.

I removed an outdated variable 'flprint'.

I may have done something else as well.