Simplify: add manipinput.sh
This bash script accepts as arguments 1)line number of
string to change, 2)string to change from, 3)string to
change to.

Since this 'Simplify' campaign has reduce the code to
running one form of transport at a time, and it is often
useful to compare results, this bash file allows for
changes in the input file easily from the command line
for easier comparisons between output from different
inputs.