Simplify: improve manipinput.sh
I improved this file with more comments and new functionality.

It can now be used without specifying line number.
The first occurance of the replace string will be replaced.
It may still be advantageous to specify line number to ensure
desired string will be replaced in case the string occurs before
the place it is to be replaced.