Simplify: create saveinput.sh
saveinput.sh has four moded of operation.
1) no input - save a copy of the input file with a suffix
   including the date and time of the call in texts/inputlog
2) 'ls' - list the files in texts/inputlog
3) 'ws' - count the files in texts/inputlog
4) 'rm' - remove the files in texts/inputlog

This script is called as soon as a call to the main executable
passes the input tests.  At that time the current date and time
is also printed to the screen.  The point of this script
is to have a record of what input corresponds to what run.
I expect this to be especially handy when running several different
runs at the same time, especially if they are long.