LN Proc: leakage output sci notation
Leakage averages and standard deviations had been given
in floating point.  Fichtl problems are thick and yield
very small transmission values that appeared to be zero
even though they are not--if you run enough particles, that is.
'MCprintstats' now tests whether any output is below a certain
value, and if any is, prints in scientific notation instead of
floating point.

Also I think there is something wrong with LPMC, but that is
not the focus right now so I do not expect to work on it
in the near future.