Simplify: realization rngs
I handled setting random number seeds for binary realz and KL realz.

For binary realizations, realizations are reproducible for same
rngseed now implicitly.  This is needed because rng are sampled
between realization generation.

For KL realizations, realizations are reproducible for same
rngseed implicitly as well.
Realizations are sampled in a correlated way when sampling
xi values from binary and contin when abs and scat are set
to the same and inverse sampling is used on the Gaussian--
as opposed to Box-Muller sampling.  If abs and scat are
set to be different, there is correlation between the
binary or same case and the abs xs.