Simplify: 'var maintained' to 'mean sqr err maint'
The mean square error throughout the slab is what I was
calculating anyway.  For normalized eigenvalues, this value
is equal to the length of the slab.  This allows a much more
simple calculation that does not require a convergence study.
I therefore removed the convergence study and just calculated
the quantities I want the easy way.  This allowed the removal
of two inputs, and significant simplification in KL_Eigenvalue.