LN Proc: LN transform for scatrat lims
I added code to set appropriate values for xs aves and vars
for abs and scattering through the log-normal transformation
when the scattering ratio is 0.0 or 1.0.  The variance becauses
zero and the xs becomes negative infinity, which I have
approximated with a negative number of large magnitude.

I also formatted printing in the time module.