LN Proc: leakage pdf print vals
As well as letting the code plot leakage pdfs, I also
print the values to a file and store this file in a directory
in the 'auxiliary' folder.  I plan to add a python script which
can read this data and create a pdf and associated plot with the
bounds and bin number the user choses.  This allows binning the
way you want, and being able to play with different binnings, without
having to re-run the code--something I expect to be very useful
in creating a benchmark run.