LN Proc: s-anumEigs pt1
This is the first step in allowing the scattering and absorption
cross sections to use a different number of eigenmodes in the KL
expansion.

If the user puts zero for the number of eigenmodes, the code will
use the individual scattering and absorption numbers of eigenmodes
and set 'numEigs' equal to the max of these two.  numEigs is still
used to solve things like coefficients.

If the user puts a number for numEigs, 'snumEigs' and 'anumEigs'
will each be set to 'numEigs'.  This re-creates what the code has
done before this commit.

Here I have added 'snumEigs' and 'anumEigs', and set the values
of the three variables mentioned here according to the scheme
described here.  New variables have not been utilized yet.