LN Proc: SC, quad orders input
I read 'Qtemp'.  I read 'Qs' values where Qs is allocated to
snumEigs plut anumEigs.  It is thus allocated even if these values
are not going to be used.  The right number of values must be
provided for Qs even if they are not used or the input reader
will throw an error.

If Qtemp is equal to 0, all Qs values are set to this.  Otherwise
the values read for Qs are kept.