LN Proc: SC, wgts used in stats
The weights are now used in calculations for flux values and
leakage values.  Only the required number of realizations are run.
SEM is not printed for SC routines, since it does not exist.

Results look reasonable in the case currently in the input file,
but need benchmarked.

I had started to treat 'MC' and 'SC' in similar ways by giving
the 'MC' UQ approach weights, equal to 1 divided by numRealz,
so that I could then re-use statistical routines.  This is what
I did in the python script that does the same type of UQ logic
as here.  I decided not to do this though since it would require
changing a fair amount of code I already believe works for not
much benefit.  I undid the first step of this process.